+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser/EventArguments.kt
@@ -197,6 +197,7 @@ internal object EventArguments {
         members = listOf(member),
         memberCount = 1,
         config = config,
+        lastMessageAt = date,
     )
 
     private val channelInfo = ChannelInfo(